Handle subsampling as long integers instead of 32 bits integers. This is needed because small overviews of large images require a subsampling which is about as large as the image size.
https://issues.apache.org/jira/browse/SIS-605

+++ endorsed/src/org.apache.sis.feature/main/org/apache/sis/coverage/privy/RangeArgument.java
@@ -57,7 +57,7 @@ public final class RangeArgument {
      * the {@code insertFoo(…)} methods are invoked.
      *
      * @see #insertBandDimension(GridExtent, int)
-     * @see #insertSubsampling(int[], int)
+     * @see #insertSubsampling(long[], int)
      */
     private int first, last, interval;
 
@@ -140,7 +140,7 @@ public final class RangeArgument {
 
     /**
      * Returns {@code true} if user specified all bands in increasing order.
-     * This method always return {@code false} if {@link #insertSubsampling(int[], int)} has been invoked.
+     * This method always return {@code false} if {@link #insertSubsampling(long[], int)} has been invoked.
      *
      * @return whether user specified all bands in increasing order without subsampling inserted.
      */
@@ -221,7 +221,7 @@ public final class RangeArgument {
     /**
      * Returns the i<sup>th</sup> index of the band to read from the resource, after subsampling has been applied.
      * The subsampling results from calls to {@link #insertBandDimension(GridExtent, int)} and
-     * {@link #insertSubsampling(int[], int)} methods.
+     * {@link #insertSubsampling(long[], int)} methods.
      *
      * {@snippet lang="java" :
      *     areaOfInterest = rangeIndices.insertBandDimension(areaOfInterest, bandDimension);
@@ -256,7 +256,7 @@ public final class RangeArgument {
     /**
      * Returns the given extent with a new dimension added for the bands. The extent in the new dimension
      * will range from the minimum {@code range} value to the maximum {@code range} value inclusive.
-     * This method should be used together with {@link #insertSubsampling(int[], int)}.
+     * This method should be used together with {@link #insertSubsampling(long[], int)}.
      *
      * <h4>Use case</h4>
      * This method is useful for reading a <var>n</var>-dimensional data cube with values stored in a
@@ -294,7 +294,7 @@ public final class RangeArgument {
      * @param  bandDimension  index of the band dimension.
      * @return a new subsampling array with the same values as the given array plus one dimension for bands.
      */
-    public int[] insertSubsampling(int[] subsampling, final int bandDimension) {
+    public long[] insertSubsampling(long[] subsampling, final int bandDimension) {
         final int[] delta = new int[packed.length - 1];
         for (int i=0; i<delta.length; i++) {
             delta[i] = getSourceIndex(i+1) - getSourceIndex(i);
